chore(rust): move `BindingBuiltinPluginName` to a standalone file
