+++ crates/rolldown_binding/src/options/plugin/types/mod.rs
@@ -1,4 +1,5 @@
 pub mod binding_asset_source;
+pub mod binding_builtin_plugin_name;
 pub mod binding_emitted_asset;
 pub mod binding_hook_filter;
 pub mod binding_hook_load_output;
