+++ packages/rolldown/src/rolldown-binding.wasi.cjs
@@ -117,19 +117,19 @@ function __napi_rs_initialize_modules(__napiInstance) {
   __napiInstance.exports['__napi_register__BindingTransformPluginContext_struct_34']?.()
   __napiInstance.exports['__napi_register__BindingTransformPluginContext_impl_36']?.()
   __napiInstance.exports['__napi_register__BindingAssetSource_struct_37']?.()
-  __napiInstance.exports['__napi_register__BindingEmittedAsset_struct_38']?.()
-  __napiInstance.exports['__napi_register__BindingGeneralHookFilter_struct_39']?.()
-  __napiInstance.exports['__napi_register__BindingTransformHookFilter_struct_40']?.()
-  __napiInstance.exports['__napi_register__BindingHookLoadOutput_struct_41']?.()
-  __napiInstance.exports['__napi_register__BindingHookRenderChunkOutput_struct_42']?.()
-  __napiInstance.exports['__napi_register__BindingHookResolveIdExtraArgs_struct_43']?.()
-  __napiInstance.exports['__napi_register__BindingHookResolveIdOutput_struct_44']?.()
-  __napiInstance.exports['__napi_register__BindingHookSideEffects_45']?.()
-  __napiInstance.exports['__napi_register__BindingHookTransformOutput_struct_46']?.()
-  __napiInstance.exports['__napi_register__BindingPluginContextResolveOptions_struct_47']?.()
-  __napiInstance.exports['__napi_register__BindingTransformHookExtraArgs_struct_48']?.()
-  __napiInstance.exports['__napi_register__BindingBuiltinPlugin_struct_49']?.()
-  __napiInstance.exports['__napi_register__BindingBuiltinPluginName_50']?.()
+  __napiInstance.exports['__napi_register__BindingBuiltinPluginName_38']?.()
+  __napiInstance.exports['__napi_register__BindingEmittedAsset_struct_39']?.()
+  __napiInstance.exports['__napi_register__BindingGeneralHookFilter_struct_40']?.()
+  __napiInstance.exports['__napi_register__BindingTransformHookFilter_struct_41']?.()
+  __napiInstance.exports['__napi_register__BindingHookLoadOutput_struct_42']?.()
+  __napiInstance.exports['__napi_register__BindingHookRenderChunkOutput_struct_43']?.()
+  __napiInstance.exports['__napi_register__BindingHookResolveIdExtraArgs_struct_44']?.()
+  __napiInstance.exports['__napi_register__BindingHookResolveIdOutput_struct_45']?.()
+  __napiInstance.exports['__napi_register__BindingHookSideEffects_46']?.()
+  __napiInstance.exports['__napi_register__BindingHookTransformOutput_struct_47']?.()
+  __napiInstance.exports['__napi_register__BindingPluginContextResolveOptions_struct_48']?.()
+  __napiInstance.exports['__napi_register__BindingTransformHookExtraArgs_struct_49']?.()
+  __napiInstance.exports['__napi_register__BindingBuiltinPlugin_struct_50']?.()
   __napiInstance.exports['__napi_register__BindingGlobImportPluginConfig_struct_51']?.()
   __napiInstance.exports['__napi_register__BindingManifestPluginConfig_struct_52']?.()
   __napiInstance.exports['__napi_register__BindingModulePreloadPolyfillPluginConfig_struct_53']?.()
